Add some test coverage for the simplified source dependency mapping DSL.
