+++ subprojects/version-control/src/integTest/groovy/org/gradle/vcs/internal/DeclarativeSourceDependencyMultiprojectIntegrationTest.groovy
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.vcs.internal
+
+import org.gradle.vcs.fixtures.GitRepository
+
+class DeclarativeSourceDependencyMultiprojectIntegrationTest extends AbstractSourceDependencyMultiprojectIntegrationTest {
+    void mappingFor(GitRepository gitRepo = repo, String... coords) {
+        settingsFile << """
+            sourceControl {
+                gitRepository("${gitRepo.url}") {
+        """
+        coords.each { coord ->
+            settingsFile << """
+                    producesModule("${coord}")
+            """
+        }
+        settingsFile << """
+                }
+            }
+        """
+    }
+}
