+++ subprojects/version-control/src/integTest/groovy/org/gradle/vcs/internal/MappingSourceDependencyIntegrationTest.groovy
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.vcs.internal
+
+import org.gradle.vcs.fixtures.GitRepository
+import org.gradle.vcs.git.GitVersionControlSpec
+
+
+class MappingSourceDependencyIntegrationTest extends AbstractSourceDependencyMultiprojectIntegrationTest {
+    void mappingFor(GitRepository gitRepo = repo, String... coords) {
+        settingsFile << """
+            sourceControl {
+                vcsMappings {
+        """
+        coords.each { coord ->
+            settingsFile << """
+                    withModule("${coord}") {
+                        from(${GitVersionControlSpec.name}) {
+                            url = uri("$gitRepo.url")
+                        }
+                    }
+            """
+        }
+        settingsFile << """
+                }
+            }
+        """
+    }
+}
